[GOHW-3]: реализация взаимодействия с бд добавлена. Добавлен pgbouncer

See merge request sirejja/Homework!4

+++ loms/internal/api/v1/stocks_handler.go
@@ -28,7 +28,7 @@ func (s *Implementation) Stocks(ctx context.Context, req *desc.StocksRequest) (*
 
 	stocksResp := make([]*desc.Stock, 0, len(stocks))
 	for _, stock := range stocks {
-		stocksResp = append(stocksResp, &desc.Stock{Count: stock.Count, WarehouseID: stock.WarehouseID})
+		stocksResp = append(stocksResp, &desc.Stock{Count: uint64(stock.Count), WarehouseID: int64(stock.WarehouseID)})
 	}
 
 	return &desc.StocksResponse{Stocks: stocksResp}, nil
